fix bug image review, refactor code

+++ server/filesharing-filehandler/src/main/java/org/example/filesharing/entities/models/annotation/AnnotationTimeCode.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class AnnotationTimeCode {
-    private Long startMs;
-    private Long endMs;
+    private Double startMs;
+    private Double endMs;
 }
